requirements.txt
groq
langchain
langchain_community
chromadb
dotenv
gtts
googletrans
python-telegram-bot
